Fix compiler complaining about nullable
Dumb compiler doesn't know that some methods will only be used after the method that creates the game object has been created. it would be really cool if that was possible bit unfortunately it isn't possible. It would probably require some wierd analysis that simply isn't possibl condiering the jit nature of the compiler. Suck but that's just the way it's. Would be noce though. Maybe somekind og support for a state machine in C# that makes it easy to enable/disable methods under certain conditions